move fdrexperimental.py to fdr_/classifiers.py, move out plots
