+++ tests/performance_tests/fdr_test.py
@@ -13,11 +13,12 @@ import torch
 from alphabase.tools.data_downloader import DataShareDownloader
 
 import alphadia
-from alphadia import fdr, fdrexperimental
+from alphadia import fdr
+from alphadia.fdr_.classifiers import BinaryClassifierLegacyNewBatching
 from alphadia.workflow.peptidecentric.peptidecentric import feature_columns
 
 classifiers = {
-    "legacy_new_batching": fdrexperimental.BinaryClassifierLegacyNewBatching,
+    "legacy_new_batching": BinaryClassifierLegacyNewBatching,
 }
 
 parser = argparse.ArgumentParser(description="Run the fdr test.")
